+++ HISTORY.rst
@@ -8,9 +8,13 @@ Unreleased
 
 **Features**
 
+- New lyrics: CSV loader from path or stream
+
 **Bugfixes**
 
 **Improvements**
+
+- Add tests for the existing lyrics
 
 **Documentation**